Edit transaction refactor
Refactored edit transaction form to using the TextEditingController for listening to changes. Replaces onChanged function. The controller also has a listener for performing additional operations. In this case, the _onAmountChanged listener is in charge of parsing the amount to a double when form is submitted with a new value. Is it important to dispose of controllers and listeners in order to utilize app resources effectively and improve performance.